move the lip operator to operator module
