+++ src/benchmark/ssb_workload_lip.cc
@@ -21,9 +21,9 @@
 #include "operators/fused/select_build_hash.h"
 #include "operators/join.h"
 #include "operators/join_graph.h"
+#include "operators/lip.h"
 #include "operators/predicate.h"
 #include "operators/select.h"
-#include "operators/utils/lip.h"
 #include "scheduler/scheduler.h"
 #include "ssb_workload.h"
 #include "storage/util.h"
