+++ src/operators/lip.h
